Fixed: premiumize.me, my bad! - re: IRC
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@21218 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/hoster/PremiumizeMe.java
@@ -228,7 +228,7 @@ public class PremiumizeMe extends PluginForHost {
         }
         String trafficleft_bytes = br.getRegex("trafficleft_bytes\":(-?[\\d\\.]+)").getMatch(0);
         if (trafficleft_bytes != null) {
-            if (Integer.parseInt(trafficleft_bytes) <= 0) trafficleft_bytes = "0";
+            if (Long.parseLong(trafficleft_bytes) <= 0) trafficleft_bytes = "0";
             ai.setTrafficLeft(trafficleft_bytes);
         } else {
             ai.setUnlimitedTraffic();
